+++ R/get_marginalmeans.R
@@ -547,12 +547,6 @@ get_marginalmeans <- function(
   verbose = TRUE,
   ...
 ) {
-  # special case: calculating context effects for models with within-between
-  # effects. In this case, we don't want any further checks
-  if (identical(comparison, "context") || identical(comparison, "context_pairwise")) {
-    return(list(by = by, contrast = contrast))
-  }
-
   # Gather info and data from model
   model_data <- insight::get_data(model, verbose = FALSE)