+++ tests/testthat/test-estimate_contrasts_context.R
@@ -12,11 +12,7 @@ test_that("estimate_contrast, context effects, linear", {
   b <- coef(summary(m))[2:3, 1]
   se <- coef(summary(m))[2:3, 2]
 
-  out <- estimate_contrasts(
-    m,
-    c("bill_len_between", "bill_len_within"),
-    comparison = "context"
-  )
+  out <- estimate_contrasts(m, c("bill_len_between", "bill_len_within"))
   expect_equal(out$Difference, b[2] - b[1], tolerance = 1e-4, ignore_attr = TRUE)
   expect_equal(out$SE, sqrt((se[1]^2 + se[2]^2)), tolerance = 1e-4, ignore_attr = TRUE)
   expect_true(!is.null(out$p))
@@ -24,24 +20,23 @@ test_that("estimate_contrast, context effects, linear", {
   output <- capture.output(out)
   expect_identical(output[3], "Difference |   SE |       95% CI |     z |      p")
 
-  m <- lm(bill_dep ~ year * (bill_len_between + bill_len_within), data = d)
-  out <- estimate_contrasts(
-    m,
-    c("bill_len_between", "bill_len_within"),
-    by = "year",
-    comparison = "context"
+  expect_message(
+    estimate_contrasts(
+      m,
+      c("bill_len_between", "bill_len_within"),
+      comparison = "difference"
+    ),
+    regex = "The `comparison` argument will be set"
   )
+
+  m <- lm(bill_dep ~ year * (bill_len_between + bill_len_within), data = d)
+  out <- estimate_contrasts(m, c("bill_len_between", "bill_len_within"), by = "year")
   expect_named(out, c("year", "Difference", "SE", "CI_low", "CI_high", "z", "p"))
   x1 <- estimate_slopes(m, "bill_len_within", by = "year")
   x2 <- estimate_slopes(m, "bill_len_between", by = "year")
   expect_equal(out$Difference, x1$Slope - x2$Slope, tolerance = 1e-4)
 
-  out <- estimate_contrasts(
-    m,
-    c("bill_len_between", "bill_len_within"),
-    by = "year",
-    comparison = "context_pairwise"
-  )
+  out <- estimate_contrasts(m, c("bill_len_between", "bill_len_within", "year"))
   expect_named(
     out,
     c("Level1", "Level2", "Parameter", "Difference", "SE", "CI_low", "CI_high", "z", "p")
@@ -63,11 +58,7 @@ test_that("estimate_contrast, context effects, glm", {
   b <- coef(summary(m))[2:3, 1]
   se <- coef(summary(m))[2:3, 2]
 
-  out <- estimate_contrasts(
-    m,
-    c("bill_len_between", "bill_len_within"),
-    comparison = "context"
-  )
+  out <- estimate_contrasts(m, c("bill_len_between", "bill_len_within"))
   expect_equal(out$Odds_Ratio, exp(b[2] - b[1]), tolerance = 1e-4, ignore_attr = TRUE)
   expect_true(!is.null(out$p))
 
@@ -77,15 +68,6 @@ test_that("estimate_contrast, context effects, glm", {
   out <- estimate_contrasts(
     m,
     c("bill_len_between", "bill_len_within"),
-    comparison = "slope"
-  )
-  expect_equal(out$Odds_Ratio, exp(b[2] - b[1]), tolerance = 1e-4, ignore_attr = TRUE)
-  expect_true(!is.null(out$p))
-
-  out <- estimate_contrasts(
-    m,
-    c("bill_len_between", "bill_len_within"),
-    comparison = "context",
     predict = "response"
   )
   expect_equal(out$Probability, -0.01784138, tolerance = 1e-4, ignore_attr = TRUE)